Fill in "void" data with surrounding data.

+++ src/Lib/Array/array.hxx
@@ -67,7 +67,7 @@ private:
     double col_step, row_step;
     
     // pointers to the actual grid data allocated here
-    float (*in_data)[ARRAY_SIZE_1];
+    int (*in_data)[ARRAY_SIZE_1];
     // float (*out_data)[ARRAY_SIZE_1];
 
     // output nodes
